Unify devlog managers and update tests
